reformat torchdata
