+++ opsml/model/interfaces/backups.py
@@ -9,7 +9,9 @@ class HuggingFaceModelNoModule(ModelInterface):
     @model_validator(mode="before")
     @classmethod
     def check_model(cls, model_args: Dict[str, Any]) -> Dict[str, Any]:
-        raise ModuleNotFoundError("HuggingFaceModel requires transformers to be installed. Please install transformers.")
+        raise ModuleNotFoundError(
+            "HuggingFaceModel requires transformers to be installed. Please install transformers."
+        )
 
     @staticmethod
     def name() -> str:
@@ -31,7 +33,9 @@ class LightningModelNoModule(ModelInterface):
     @model_validator(mode="before")
     @classmethod
     def check_model(cls, model_args: Dict[str, Any]) -> Dict[str, Any]:
-        raise ModuleNotFoundError("LightningModel requires pytorch lightning to be installed. Please install lightning.")
+        raise ModuleNotFoundError(
+            "LightningModel requires pytorch lightning to be installed. Please install lightning."
+        )
 
     @staticmethod
     def name() -> str:
