+++ opsml/model/interfaces/pytorch_lightning.py
@@ -112,7 +112,9 @@ try:
             try:
                 if model_arch is not None:
                     # attempt to load checkpoint into model
-                    assert issubclass(model_arch, LightningModule), "Model architecture must be a subclass of LightningModule"
+                    assert issubclass(
+                        model_arch, LightningModule
+                    ), "Model architecture must be a subclass of LightningModule"
                     self.model = model_arch.load_from_checkpoint(checkpoint_path=path, **kwargs)
 
                 else:
